requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.122.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dotenv==1.2.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
